refactor(plot): move trial phase drawing to utils_common
Relocate PHASE_COLORS, get_trial_phase_boundaries, and draw_trial_phases
from plot_utils to utils_common.utils so plant_plotting can reuse them.
Adds time_unit_s parameter for seconds-scale axes.
